+++ packages/storybook-new/.storybook/manager.js
@@ -1,0 +1,25 @@
+import { addons } from '@storybook/addons';
+
+addons.setConfig({
+  isFullscreen: false,
+  showNav: true,
+  showPanel: true,
+  panelPosition: 'bottom',
+  enableShortcuts: true,
+  showToolbar: true,
+  theme: undefined,
+  selectedPanel: undefined,
+  initialActive: 'sidebar',
+  sidebar: {
+    showRoots: false,
+    collapsedRoots: ['other'],
+  },
+  toolbar: {
+    title: { hidden: false },
+    zoom: { hidden: true },
+    eject: { hidden: false },
+    copy: { hidden: false },
+    fullscreen: { hidden: true },
+    'storybook/background': { hidden: true },
+  },
+});