+++ packages/tree-view/src/behaviors/TreeViewBehavior.js
@@ -113,24 +113,21 @@ const TreeViewBehavior = (props) => {
 
     const domNodeList = treeViewRef.current.querySelectorAll("li");
     const treeItemArrayControl =
-      getTreeItemArray().length !== domNodeList.length || props.treeNode
+      getTreeItemArray()?.length !== domNodeList?.length || props.treeNode
         ? buildTreeItemIdArray(Array.prototype.slice.call(domNodeList))
         : getTreeItemArray();
 
     const lowerLimit = 0;
     const upperLimit = treeItemArrayControl.length - 1;
 
-    switch (event.keyCode) {
-      // Arrow Down
-      case 40: {
+    switch (event.code) {
+      case "ArrowDown": {
         event.preventDefault();
         if (getActiveTreeItemIndex() === upperLimit) {
           setActiveTreeItemIndex(lowerLimit);
-
           checkScroll(treeItemArrayControl[lowerLimit], treeViewRef);
         } else {
           setActiveTreeItemIndex(getActiveTreeItemIndex() + 1);
-
           checkScroll(
             treeItemArrayControl[getActiveTreeItemIndex() + 1],
             treeViewRef
@@ -138,9 +135,7 @@ const TreeViewBehavior = (props) => {
         }
         break;
       }
-
-      // Arrow Up
-      case 38: {
+      case "ArrowUp": {
         event.preventDefault();
         if (getActiveTreeItemIndex() <= lowerLimit) {
           setActiveTreeItemIndex(upperLimit);
@@ -156,15 +151,12 @@ const TreeViewBehavior = (props) => {
         }
         break;
       }
-
-      // Enter
-      case 13: {
+      case "Enter": {
         event.preventDefault();
         setKeyboardOpenId(getActiveTreeItemId());
         break;
       }
-      // Space
-      case 32: {
+      case "Space": {
         event.preventDefault();
         if (isControlled()) {
           return;
@@ -172,7 +164,6 @@ const TreeViewBehavior = (props) => {
         setActiveTreeItemId(getActiveTreeItemId());
         break;
       }
-
       default:
     }
   };